Revert "Add performance test for set/delete routes (#2344)"
This reverts commit 814f702b26ba8c5bf790506feb1dda705df7676b.
